refactor!: rename text module to textual
BREAKING CHANGE: Adjective usage is more proper than noun usage, like
that we are using numeric for the module that deals with numbers.
